feat: preserve workout history when deleting exercises

+++ server/controllers/exerciseController.js
@@ -1,22 +1,42 @@
 const Exercise = require("../models/exerciseModel");
 const Category = require("../models/categoryModel");
+const Workout = require("../models/workoutModel");
 
 const getExercises = async (req, res) => {
   const user_id = req.user._id;
 
   const exercises = await Exercise
-    .find({ user_id })
+    .find({
+      user_id,
+      isArchived: false
+    })
     .populate("categoryId")
     .sort({ name: 1 });
 
-  res.status(200).json(exercises);
+  const exercisesWithCounts = await Promise.all(
+    exercises.map(async exercise => {
+      const workoutCount = await Workout.countDocuments({
+        user_id,
+        "exercises.exerciseId": exercise._id
+      });
+
+      return {
+        ...exercise.toObject(),
+        workoutCount
+      };
+    })
+  );
+
+  res.status(200).json(exercisesWithCounts);
 };
 
 const createExercise = async (req, res) => {
   const { name, categoryId } = req.body;
   const user_id = req.user._id;
 
-  if (!name) {
+  const normalizedName = name.trim();
+
+  if (!normalizedName) {
     return res.status(400).json({
       error: "Exercise name is required"
     });
@@ -31,7 +51,7 @@ const createExercise = async (req, res) => {
   const existingExercise = await Exercise.findOne({
     user_id,
     name: {
-      $regex: `^${name}$`,
+      $regex: `^${normalizedName}$`,
       $options: "i"
     }
   });
@@ -44,6 +64,7 @@ const createExercise = async (req, res) => {
 
   const exercise = await Exercise.create({
     ...req.body,
+    name: normalizedName,
     user_id
   });
 
@@ -51,7 +72,10 @@ const createExercise = async (req, res) => {
     .findById(exercise._id)
     .populate("categoryId");
 
-  res.status(201).json(populatedExercise);
+  res.status(201).json({
+    ...populatedExercise.toObject(),
+    workoutCount: 0
+  });
 };
 
 const updateExercise = async (req, res) => {
@@ -112,16 +136,51 @@ const updateExercise = async (req, res) => {
     .findById(exercise._id)
     .populate("categoryId");
 
-  res.status(200).json(populatedExercise);
+  const workoutCount = await Workout.countDocuments({
+    user_id,
+    "exercises.exerciseId": exercise._id
+  });
+
+  res.status(200).json({
+    ...populatedExercise.toObject(),
+    workoutCount
+  });
 };
 
 const deleteExercise = async (req, res) => {
   const user_id = req.user._id;
   const { id } = req.params;
 
-  const exercise = await Exercise.findOneAndDelete(
-    { _id: id, user_id }
-  );
+  const workoutCount = await Workout.countDocuments({
+    user_id,
+    "exercises.exerciseId": id
+  });
+
+  let exercise;
+
+  if (workoutCount === 0) {
+
+    exercise = await Exercise.findOneAndDelete({
+      _id: id,
+      user_id
+    });
+
+  } else {
+
+    exercise = await Exercise.findOneAndUpdate(
+      {
+        _id: id,
+        user_id
+      },
+      {
+        isArchived: true
+      },
+      {
+        new: true
+      }
+    );
+
+  }
 
   if (!exercise) {
     return res.status(400).json({
@@ -129,7 +188,15 @@ const deleteExercise = async (req, res) => {
     });
   }
 
-  res.status(200).json(exercise);
+  const populatedExercise = await Exercise
+    .findById(exercise._id)
+    .populate("categoryId");
+
+  res.status(200).json({
+    exercise: populatedExercise,
+    workoutCount,
+    archived: workoutCount > 0
+  });
 };
 
 module.exports = { getExercises, createExercise, updateExercise, deleteExercise };
